Fixes segmentation fault that only happens on export
It seems that when the box is removed from memory, the player is still colliding with it, leading to a segmentation fault. This PR makes it so that when the timer ticks, the boolean variable "destroy_block" is being set to true, and on the next physics step, if destroy_block is true, the node is being removed from the SceneTree. On the new Box.gd script, the "_exit_tree()" method is executed when the node is removed from the scene tree, and it calls "queue_free()", which frees the node from memory, preventing a memory leak.

This PR also fixes a memory leak caused by instantiating a box node on the timer timeout, and never adding it to the scene tree, thus it remains as an orphan node. This instantiation is not needed, since it's already being done in the "add_box()" method.